Initial mergin of decorations to package
