+++ src/modeltools/ModelTools.jl
@@ -162,6 +162,7 @@ include("SimpleModels.jl")
 include("ExpODEModels.jl")
 include("ExpStateModels.jl")
 include("MonomialRegressionModels.jl")
+include("Decorations.jl")
 include("PetriModels.jl")
 include("OpenModels.jl")
 include("OpenPetris.jl")
